+++ README.md
@@ -3,13 +3,13 @@
   <p>Adalah sebuah Website untuk mendapatkan jadwal sholat, puasa sunnah, dan baca Al-Qur'an</p>
 </div>
 
-## 🗒️Notes ️
+## 🗒️ Notes ️
 
 - Ini adalah branch `dev`, dimana semua perubahan atau fitur baru bakal dites terlebih dahulu disini. Jika bekerja sesuai ekspektasi, maka bakal dibuat PR dan dimerge ke branch `master`.
 - Untuk penggunaan pencarian Surah dan Kota, kamu bisa memanfaatkan shortcut keyboard `Shift` + `Enter` untuk langsung focus ke search bar
 - Untuk menghilangkan focus dari search bar, cukup manfaatkan shortcut `Esc`
 
-## ⚡TODO
+## ⚡ TODO
 
 - Detect kombinasi shortcut key buat search yang lebih cepet
 - Add kbd tag in search input
