+++ src/components/templates/layout/index.tsx
@@ -7,7 +7,7 @@ const Layout = ({ children, title }: Children) => {
       <Head>
         <title>{title}</title>
       </Head>
-      <section className="flex w-full max-w-full flex-col items-center justify-center bg-gray-50 px-5 pt-8 pb-24 text-center dark:bg-gray-900 md:px-8">
+      <section className="flex min-h-screen w-full max-w-full flex-col items-center bg-gray-50 px-5 pt-8 pb-24 text-center dark:bg-gray-900 md:px-8">
         <div className="flex w-full max-w-full flex-col items-center justify-center gap-7">
           {children}
         </div>
